requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
antlr4-python3-runtime=4.9.3=pypi_0
ca-certificates=2023.12.12=h06a4308_0
certifi=2022.12.7=py37h06a4308_0
charset-normalizer=3.3.2=pypi_0
clip=1.0=pypi_0
ftfy=6.1.1=pypi_0
hydra-core=1.3.2=pypi_0
idna=3.6=pypi_0
importlib-resources=5.12.0=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.3.2=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
markupsafe=2.1.3=pypi_0
ncurses=6.4=h6a678d5_0
numpy=1.21.6=pypi_0
omegaconf=2.3.0=pypi_0
openssl=1.1.1w=h7f8727e_0
packaging=23.2=pypi_0
pillow=9.5.0=pypi_0
pip=22.3.1=py37h06a4308_0
psutil=5.9.7=pypi_0
pyparsing=3.1.1=pypi_0
python=3.7.16=h7a1cb2a_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
regex=2023.10.3=pypi_0
requests=2.31.0=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=pypi_0
setuptools=65.6.3=py37h06a4308_0
sqlite=3.41.2=h5eee18b_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h1ccaba5_0
torch=1.13.1+cu117=pypi_0
torch-geometric=2.3.1=pypi_0
torchaudio=0.13.1+cu117=pypi_0
torchvision=0.14.1+cu117=pypi_0
tqdm=4.66.1=pypi_0
typing-extensions=4.7.1=pypi_0
urllib3=2.0.7=pypi_0
wcwidth=0.2.12=pypi_0
wheel=0.38.4=py37h06a4308_0
xz=5.4.5=h5eee18b_0
zipp=3.15.0=pypi_0
zlib=1.2.13=h5eee18b_0
